fix bug
ya estuvo
